crud producto edit ferney develop

+++ src/main/java/com/example/subastasquindio/mapping/dto/CompradorDto.java
@@ -1,12 +1,16 @@
 package com.example.subastasquindio.mapping.dto;
 
+import com.example.subastasquindio.model.Puja;
 import com.example.subastasquindio.model.Usuario;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public record CompradorDto (
         String nombre,
         String apellido,
         String cedula,
-        int edad,
+        Double edad,
         String telefono,
         String direccion){
